Change ollama configuration

+++ llm/src/main/kotlin/ru/psychologicalTesting/llm/config/ollama/OllamaConfig.kt
@@ -9,5 +9,5 @@ interface OllamaConfig {
     val chatSystemPrompt: String
     val testTranscriptionSystemPrompt: String
     val testTranscriptionTemperature: Double
-    val testTranscriptionMaxOutputTokens: Int
+    val testTranscriptionMaxOutputTokens: Long
 }
